refactor: align exam and notification identifiers to long

+++ EduOS.Core/Interfaces/IRepositories/IMarkEntryRepository.cs
@@ -4,9 +4,9 @@ namespace EduOS.Core.Interfaces.IRepositories
 {
     public interface IMarkEntryRepository : IGenericRepository<MarkEntry>
     {
-        Task<List<MarkEntry>> GetByExamAndStudentAsync(int examId, int studentId);
-        Task<List<MarkEntry>> GetByExamAndSubjectAsync(int examId, int subjectId, int classId);
-        Task<MarkEntry?> GetExistingAsync(int examId, int studentId, int subjectId);
-        Task<bool> IsAllMarksEnteredAsync(int examId, int classId);
+        Task<List<MarkEntry>> GetByExamAndStudentAsync(long examId, long studentId);
+        Task<List<MarkEntry>> GetByExamAndSubjectAsync(long examId, long subjectId, long classId);
+        Task<MarkEntry?> GetExistingAsync(long examId, long studentId, long subjectId);
+        Task<bool> IsAllMarksEnteredAsync(long examId, long classId);
     }
 }
